Move dictionary to a crate
To simplify the main code and make the library interoperable, I've moved
the dictionary and it's types to a separate crate.
