+++ rn_dictionary/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "rn_dictionary"
+version = "0.1.0"
+authors = ["Kevin Choubacha <kchoubacha@weedmaps.com>"]
+
+[dependencies]
+Inflector = "0.11"
