+++ rn_dictionary/src/adjectives.rs
