+++ rn_dictionary/src/adverbs.rs
