+++ rn_dictionary/src/sha_part.rs
@@ -1,5 +1,5 @@
 use std::str::FromStr;
-use dictionary::sha_result::{ParseShaError, ShaResult};
+use sha_result::{ParseShaError, ShaResult};
 
 /// A sha part. This is a slice of a sha that corresponds that can generally correspond
 /// to a word. The hash is the usize equivalent that will then be used to lookup
