+++ rn_dictionary/src/sha_result.rs
